% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_allele_frequency}
\alias{add_allele_frequency}
\title{Add within-sample allele frequency}
\usage{
add_allele_frequency(allele_data, group_cols = c("sample_name", "target_name"))
}
\arguments{
\item{allele_data}{An allele table with \code{reads}.}

\item{group_cols}{Columns identifying a locus within a sample.}
}
\value{
\code{allele_data} with an \code{AlleleFreq} column.
}
\description{
Add within-sample allele frequency
}
