% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{allele_table_group_cols}
\alias{allele_table_group_cols}
\title{Locus columns for Mad4hatter output tables}
\usage{
allele_table_group_cols()
}
\value{
A named list of the columns identifying a locus within a sample for
each allele table Mad4hatter produces.
}
\description{
Locus columns for Mad4hatter output tables
}
