% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{count_targets_per_reaction}
\alias{count_targets_per_reaction}
\title{Count targets per reaction}
\usage{
count_targets_per_reaction(panel_reactions)
}
\arguments{
\item{panel_reactions}{Panel information with reactions, from
\code{\link[=assign_reactions]{assign_reactions()}}.}
}
\value{
A data frame with \code{reaction} and \code{nreactionloci}.
}
\description{
Count targets per reaction
}
