% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_tabset}
\alias{create_tabset}
\title{Print plots as a Quarto tabset}
\usage{
create_tabset(plots)
}
\arguments{
\item{plots}{Named list of ggplots; names are used as tab titles.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Must be called from a chunk with \code{results: asis}.
}
