% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_balancing_plot}
\alias{generate_balancing_plot}
\title{Balancing across batches}
\usage{
generate_balancing_plot(
  summary_samples,
  sample_colours = default_sample_colours
)
}
\arguments{
\item{summary_samples}{Output of \code{\link[=summarise_samples]{summarise_samples()}}.}

\item{sample_colours}{Named colours for each sample type.}
}
\value{
A ggplot of total reads per sample by batch.
}
\description{
Balancing across batches
}
