% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_dimer_plot}
\alias{generate_dimer_plot}
\title{Primer dimer content}
\usage{
generate_dimer_plot(
  sample_coverage_with_manifest,
  sample_colours = default_sample_colours
)
}
\arguments{
\item{sample_coverage_with_manifest}{Output of \code{\link[=merge_sample_coverage]{merge_sample_coverage()}}.}

\item{sample_colours}{Named colours for each sample type.}
}
\value{
A ggplot of input reads against percentage of dimers per batch.
}
\description{
Primer dimer content
}
