% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{interactive_plot}
\alias{interactive_plot}
\title{Interactive version of a ggplot}
\usage{
interactive_plot(p, ...)
}
\arguments{
\item{p}{A ggplot with a \code{text} aesthetic.}

\item{...}{Passed to \code{\link[plotly:ggplotly]{plotly::ggplotly()}}.}
}
\value{
A plotly object.
}
\description{
Interactive version of a ggplot
}
