% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{merge_amplicon_coverage}
\alias{merge_amplicon_coverage}
\title{Add the manifest and reactions to amplicon coverage}
\usage{
merge_amplicon_coverage(amplicon_coverage, manifest, panel_reactions)
}
\arguments{
\item{amplicon_coverage}{Amplicon coverage from Mad4hatter.}

\item{manifest}{Manifest from \code{\link[=read_manifest]{read_manifest()}}.}

\item{panel_reactions}{Panel information with reactions, from
\code{\link[=assign_reactions]{assign_reactions()}}.}
}
\value{
Amplicon coverage with manifest columns and a \code{reaction} column.
Targets in more than one reaction appear once per reaction.
}
\description{
Add the manifest and reactions to amplicon coverage
}
