% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{missing_targets_report}
\alias{missing_targets_report}
\title{Targets in the panel missing from the allele data}
\usage{
missing_targets_report(panel_information, allele_data)
}
\arguments{
\item{panel_information}{Panel information from Mad4hatter.}

\item{allele_data}{Allele data from Mad4hatter.}
}
\value{
\code{target_name}, \code{pool} and \code{chrom} of targets with no allele data.
}
\description{
Targets in the panel missing from the allele data
}
