% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{negative_control_amplicons}
\alias{negative_control_amplicons}
\title{Negative control amplicon coverage}
\usage{
negative_control_amplicons(amplicon_coverage_with_manifest)
}
\arguments{
\item{amplicon_coverage_with_manifest}{Output of \code{\link[=merge_amplicon_coverage]{merge_amplicon_coverage()}}.}
}
\value{
Amplicon coverage for negative controls, with a \code{negative} label
giving the batch and well.
}
\description{
Negative control amplicon coverage
}
