% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{negative_control_plot_height}
\alias{negative_control_plot_height}
\title{Figure height for negative control histograms}
\usage{
negative_control_plot_height(
  n_controls,
  ncol_layout = 3,
  base_height_per_row = 2.5,
  min_height = 6,
  max_height = 20
)
}
\arguments{
\item{n_controls}{Number of negative controls plotted.}

\item{ncol_layout}{Number of facet columns.}

\item{base_height_per_row}{Height per row of facets, in inches.}

\item{min_height, max_height}{Height limits, in inches.}
}
\value{
Figure height in inches.
}
\description{
Figure height for negative control histograms
}
