% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_negative_control_target_reads}
\alias{plot_negative_control_target_reads}
\title{Total reads per target across negative controls}
\usage{
plot_negative_control_target_reads(
  summary_data,
  reaction_colours = default_reaction_colours
)
}
\arguments{
\item{summary_data}{Output of \code{\link[=summarise_negative_control_targets]{summarise_negative_control_targets()}}.}

\item{reaction_colours}{Colours for each reaction.}
}
\value{
A ggplot with a \code{text} aesthetic for \code{\link[plotly:ggplotly]{plotly::ggplotly()}} tooltips.
}
\description{
Total reads per target across negative controls
}
