% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_positive_control_alleles}
\alias{plot_positive_control_alleles}
\title{Alleles per target in positive controls}
\usage{
plot_positive_control_alleles(category_counts)
}
\arguments{
\item{category_counts}{\code{category_counts} from
\code{\link[=summarise_positive_controls]{summarise_positive_controls()}}.}
}
\value{
A ggplot of targets per allele category for each positive control,
with a \code{text} aesthetic for \code{\link[plotly:ggplotly]{plotly::ggplotly()}} tooltips.
}
\description{
Alleles per target in positive controls
}
