% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_qc_status_by_batch}
\alias{plot_qc_status_by_batch}
\title{QC status by batch}
\usage{
plot_qc_status_by_batch(qc_summary_by_batch)
}
\arguments{
\item{qc_summary_by_batch}{Output of \code{\link[=qc_summary_by_batch]{qc_summary_by_batch()}}.}
}
\value{
A stacked bar chart of the percentage of samples in each status.
}
\description{
QC status by batch
}
