% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{qc_report_template}
\alias{qc_report_template}
\title{Path to the QC report template}
\usage{
qc_report_template()
}
\value{
Path to the Quarto template installed with the package.
}
\description{
Copy this file if you want to customise the report.
}
