% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{styled_table}
\alias{styled_table}
\title{Styled HTML table}
\usage{
styled_table(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Passed to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\value{
A \code{kableExtra} table.
}
\description{
Styled HTML table
}
